# AppKube-common-security-rbac-services/Api-Specifications/apispecs.yml
openapi: 3.0.3
info:
  title: Common security rbac services
  version: 1.0.0
  description: API for managing organizations, policies, and emails
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /policies:
    get:
      summary: Retrieve all policies
      tags:
        - Policies
      responses:
        '200':
          description: A list of policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Policy'
    post:
      summary: Create a new policy
      tags:
        - Policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        '201':
          description: The created policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
  /policies/{id}:
    get:
      summary: Retrieve a specific policy by ID
      tags:
        - Policies
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the policy to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The requested policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
    delete:
      summary: Delete a policy by ID
      tags:
        - Policies
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the policy to delete
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Policy deleted successfully
  /policies/policy-assigned-permissions:
    get:
      summary: Retrieve all policy assigned permissions
      tags:
        - Policies
      responses:
        '200':
          description: A list of policy assigned permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PolicyAssignedPermissions'
    post:
      summary: Create a new policy assigned permission
      tags:
        - Policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyAssignedPermissions'
      responses:
        '201':
          description: The created policy assigned permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyAssignedPermissions'
...
components:
  schemas:
    Organization:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        fax:
          type: string
        dateOfEstablishment:
          type: string
          format: date-time
        status:
          type: string
        cmdbOrgId:
          type: integer
          format: int64
        fileName:
          type: string
        fileStorageLocationType:
          type: string
        profileImage:
          type: string
    Policy:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        version:
          type: integer
          format: int64
        status:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PolicyAssignedPermissions'
        organization:
          $ref: '#/components/schemas/Organization'
    PolicyAssignedPermissions:
      type: object
      properties:
        id:
          type: integer
          format: int64
        policyId:
          type: integer
          format: int64
        policyName:
          type: string
        permissionCategoryId:
          type: integer
          format: int64
        permissionCategoryName:
          type: string
        permissionId:
          type: integer
          format: int64
        permissionName:
          type: string
...